https://github.com/gintool/gin/issues/36
In LocalSearch and the various utilities, where an EditType was expected
as an argument you can now specify a comma separated list of Edit
classes to include in the search; EditTypes can still be given (or a
comma separated mix of the two; e.g.
STATEMENT,gin.edit.modifynode,BinaryOperatorReplacement).

This functionality is provided by extra methods in Patch:
addRandomEditOfClass() and addRandomEditOfClasses(), and a parser method
parseEditClassesFromString() added to Edit.

NOTE: this required a change to Patch.addRandomEdit() that means the RNG is called differently. This means that although the distributions of edits sampled is similar, the precise edit sequence generated for a given RNG seed will now be different to previous versions of Gin.