Fixed a bug in Compiler and normalised classpath items in ExternalTestRunner.
The Compiler bug is due to the order of the classpath inclusion. When System.getProperty("java.class.path") has a version, say 1, of Foo.jar and the class to be compiled is from a version 2 from the Foo project, version 2 must come first in the classpath items, otherwise dependencies might not be found.